feat: refined /domain interfaces, added combat/combatState.kt

+++ domain/src/main/java/com/unibo/mobile/domain/repositories/ApiRepository.kt
@@ -6,7 +6,7 @@ import com.unibo.mobile.domain.model.entity.EnemyType
 import com.unibo.mobile.domain.model.entity.PlayerClass
 
 interface ApiRepository {
-    suspend fun getEnemies(minChallengeRating: Int, maxChallengeRating: Int): List<Enemy>
+    suspend fun getEnemies(minChallengeRating: Float, maxChallengeRating: Float): List<Enemy>
     suspend fun getAbilitiesEnemy(enemy: EnemyType): List<Ability>
     suspend fun getAbilitiesPlayerCharacter(playerClass: PlayerClass, level: Int): List<Ability>
 }
